Dynamically create the executor
Currently the executor gets created on class initialization.
This creates a problem when multiple requests are coming
at the same time. There will always be the same number of
workers in the executor which will stall the future requests
until the past requests have completed.

To address this issue, the executor will now be created
when the method gets called.